requirements.txt
PyPDF2
python-dotenv
openai
fastapi
uvicorn
python-multipart